Get rid of a lot of gcc warnings
Mostly signed/unsigned comparisons. Also some bugs, but I don't remember
where :D

+++ BBGE/AfterEffect.cpp
@@ -85,7 +85,7 @@ AfterEffectManager::~AfterEffectManager()
 
 void AfterEffectManager::deleteEffects()
 {
-	for (int i = 0; i < effects.size(); i++)
+	for (size_t i = 0; i < effects.size(); i++)
 	{
 		if (effects[i])
 		{
@@ -137,7 +137,7 @@ void AfterEffectManager::update(float dt)
 	else
 		active = false;
 
-	for (int i = 0; i < effects.size(); i++)
+	for (size_t i = 0; i < effects.size(); i++)
 	{
 		Effect *e = effects[i];
 		if (e)
